requirements.txt
python-gitlab
argparse